add 2nd use case which is create an attendance record for one student.

+++ src/main/java/edu/mum/cs/projects/attendance/repository/CourseOfferingRepository.java
@@ -10,9 +10,8 @@ import edu.mum.cs.projects.attendance.domain.entity.Course;
 import edu.mum.cs.projects.attendance.domain.entity.CourseOffering;
 
 @Repository
-public interface CourseOfferingRepository extends CrudRepository<CourseOffering, Integer> {
+public interface CourseOfferingRepository extends CrudRepository<CourseOffering, Long> {
 	List<CourseOffering> findByStartDate(Date startDate);
 	List<CourseOffering> findByCourse(Course course);
 	List<CourseOffering> findAll();
-	
 }
